moved InnerState to a private static class in FullFto
